Refactor imagenes y LandingPage
