+++ DREAM Lab Frontend/jsconfig.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+        "module": "commonjs",
+        "target": "es2016",
+        "jsx": "preserve"
+    },
+    "exclude": ["node_modules", "**/node_modules/*"]
+}
