--- DREAM Lab Frontend/src/LandingPage/LandingPageDev.jsx
@@ -1,13 +1,0 @@
-import Navbar from "../components/general/NavBar.jsx"; // Import the Navbar component
-import "./LandingPage.css";
-
-function LandingPageDev() {
-
-    return (
-        <>
-        <Navbar visible={true} view=""/>
-        </>
-    )
-}
-
-export default LandingPageDev;
