--- DREAM Lab Frontend/src/components/general/BotonCrearExperiencia.jsx
@@ -1,20 +1,0 @@
-import { useNavigate } from 'react-router-dom';
-import ExperienciaIcon from '../../images/crearExperiencia.svg'; 
-import "./BotonCrearExperiencia.css";
-
-
-function BotonCrearExperiencia() {
-    let navigate = useNavigate();
-
-    const handleClick = () => {
-        navigate(`/`); // Navega a la landing page
-    };
-
-    return (
-        <div className="flex items-center" onClick={handleClick}>
-            <img className="experienciaIcon"  src={ExperienciaIcon} alt="Crear experiencia" />
-        </div>
-    );
-}
-
-export default BotonCrearExperiencia;
