+++ DREAM Lab Frontend/src/pages/HomePage/assets/images/blob-left.png
