+++ DREAM Lab Frontend/src/pages/HomePage/assets/images/blob-right.png
