+++ DREAM Lab Frontend/src/pages/HomePage/assets/images/small-blob.png
