+++ DREAM Lab Frontend/src/pages/LandingPage/LandingPageDev.jsx
@@ -1,0 +1,12 @@
+import Navbar from "../../GlobalComponents/NavBar/NavBar.jsx"; // Import the Navbar component
+import "./LandingPage.css";
+
+function LandingPageDev() {
+    return (
+        <>
+            <Navbar visible={true} view="" />
+        </>
+    );
+}
+
+export default LandingPageDev;
