+++ DREAM Lab Frontend/src/pages/LandingPage/assets/images/astronauta-landing-page.png
