+++ DREAM Lab Frontend/src/pages/LandingPage/assets/images/blob-landing-page.png
