+++ DREAM Lab Frontend/src/pages/LandingPage/assets/images/vr-blancos-landing-page.png
